patient-information-app/reasoning/requirements.txt
fastapi==0.111.1
openai==1.36.0
uvicorn==0.30.1
python-dotenv==1.0.1
langsmith==0.1.92
langfuse==2.39.2
redis==5.0.7
